acme_blancoweb_2022-test_plus_auth/requirements.txt
Django==3.0
djangorestframework==3.11
djangorestframework-simplejwt
psycopg2
pyYAML
markdown
django-filter
drf_yasg
